+++ source/script/legacy.js
@@ -1,3 +1,3 @@
 // legacy.coffee
 
-//= include ../../bower_components/html5shiv/dist/html5shiv.js
+// = include ../../bower_components/html5shiv/dist/html5shiv.js