+++ task/default/common.js
@@ -1,13 +1,17 @@
 // common.js
 
-var gulp = require('gulp'),
-	$ = require('gulp-load-plugins')(),
-	helper = require('../helper'),
-	path = require('../path');
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var helper = require('../helper');
+var path = require('../path');
 
 // Base
 gulp.task('common', function() {
 	var name = 'Common';
+
+	if($.util.env.watch) {
+		gulp.watch(path.source.main + '*.*', ['common']);
+	}
 
 	return gulp.src(path.source.main + '*.*')
 		.pipe($.plumber(helper.error))