+++ task/default/default.js
@@ -1,8 +1,10 @@
 // default.js
 
 var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var sequence = require('run-sequence');
 
-gulp.task('default', [
-	'build',
-	'watch'
-]);
+gulp.task('default', function(callback) {
+	$.util.env.watch = true;
+	sequence(['build', 'serve'], callback);
+});