+++ task/default/fonts.js
@@ -1,10 +1,10 @@
 // fonts.js
 
-var gulp = require('gulp'),
-	$ = require('gulp-load-plugins')(),
-	config = require('../config'),
-	helper = require('../helper'),
-	path = require('../path');
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var config = require('../config');
+var helper = require('../helper');
+var path = require('../path');
 
 // Base
 gulp.task('fonts', ['fonts:common']);
@@ -12,6 +12,10 @@ gulp.task('fonts', ['fonts:common']);
 // Common
 gulp.task('fonts:common', function() {
 	var name = 'Fonts';
+
+	if($.util.env.watch) {
+		gulp.watch(path.source.font + '**/*.svg', ['fonts:common']);
+	}
 
 	return gulp.src(path.source.font + '**/*.svg')
 		.pipe($.plumber(helper.error))