+++ task/default/scripts.js
@@ -1,16 +1,36 @@
 // scripts.js
 
-var gulp = require('gulp'),
-	$ = require('gulp-load-plugins')(),
-	helper = require('../helper'),
-	path = require('../path');
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var config = require('../config');
+var helper = require('../helper');
+var path = require('../path');
 
 // Base
-gulp.task('scripts', ['scripts:common', 'scripts:vendor']);
+gulp.task('scripts', ['scripts:lint', 'scripts:common', 'scripts:vendor']);
+
+// Lint
+gulp.task('scripts:lint', function() {
+	var name = 'Lint Scripts';
+
+	if($.util.env.watch) {
+		gulp.watch([path.source.script + '**/*.js', '!' + path.source.script + 'vendor/**'], ['scripts:lint']);
+	}
+
+	return gulp.src([path.source.script + '**/*.js', '!' + path.source.script + 'vendor/**'])
+		.pipe($.plumber(helper.error))
+		.pipe($.xo(config.plugin.xo))
+		.pipe($.duration(name))
+		.pipe(helper.success(name));
+});
 
 // Common
 gulp.task('scripts:common', function() {
 	var name = 'Common Scripts';
+
+	if($.util.env.watch) {
+		gulp.watch(path.source.script + '**/*.{js,coffee}', ['scripts:common']);
+	}
 
 	return gulp.src(path.source.script + '*.{js,coffee}')
 		.pipe($.plumber(helper.error))
@@ -18,7 +38,7 @@ gulp.task('scripts:common', function() {
 		.pipe($.include())
 		.pipe($.if(/\.coffee$/, $.coffee()))
 		.pipe($.jsvalidate())
-		.pipe($.if($.util.env.production, $.uglify()))
+		.pipe($.if($.util.env.optimize, $.uglify()))
 		.pipe($.sourcemaps.write('.'))
 		.pipe(gulp.dest(path.public.script))
 		.pipe($.duration(name))
@@ -29,10 +49,14 @@ gulp.task('scripts:common', function() {
 gulp.task('scripts:vendor', function() {
 	var name = 'Vendor Scripts';
 
+	if($.util.env.watch) {
+		gulp.watch(path.source.script + 'vendor/**/*.js', ['scripts:vendor']);
+	}
+
 	return gulp.src(path.source.script + 'vendor/**/*.js')
 		.pipe($.plumber(helper.error))
 		.pipe($.include())
-		.pipe($.if($.util.env.production, $.uglify()))
+		.pipe($.if($.util.env.optimize, $.uglify()))
 		.pipe(gulp.dest(path.public.script + 'vendor/'))
 		.pipe($.duration(name))
 		.pipe(helper.success(name));