+++ task/default/styles.js
@@ -1,9 +1,9 @@
 // styles.js
 
-var gulp = require('gulp'),
-	$ = require('gulp-load-plugins')(),
-	helper = require('../helper'),
-	path = require('../path');
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var helper = require('../helper');
+var path = require('../path');
 
 // Base
 gulp.task('styles', ['styles:common']);
@@ -12,13 +12,17 @@ gulp.task('styles', ['styles:common']);
 gulp.task('styles:common', function() {
 	var name = 'Styles';
 
+	if($.util.env.watch) {
+		gulp.watch(path.source.style + '**/*.scss', ['styles:common']);
+	}
+
 	return gulp.src(path.source.style + '*.scss')
 		.pipe($.plumber(helper.error))
 		.pipe($.sourcemaps.init())
 		.pipe($.include())
 		.pipe($.sass())
 		.pipe($.autoprefixer())
-		.pipe($.if($.util.env.production, $.csso()))
+		.pipe($.if($.util.env.optimize, $.cssnano()))
 		.pipe($.sourcemaps.write('.'))
 		.pipe(gulp.dest(path.public.style))
 		.pipe($.duration(name))