+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log
 
-## [0.1.0] - 2015-04-17
+## [Next][next] 
+
+### Added
+- Support for webp images.
+- Integrate metaquery.
+- Integrate picturefill.
+- Integrate eslint via XO.
+
+### Changed
+- Swap csso for cssnano.
+- Don't log notify success messages.
+- Optimize gulp task run order.
+- Use the `watch` flag instead of a separate task.
+- Rename `production` flag to `optimize`.
+
+### Fixed
+- Include vendor.js file.
+- Use correct bower task name.
+
+## 0.1.0 - 2015-04-17
 
 ### Added
 - Lots of good stuff.
+
+[next]: https://github.com/thasmo/gulp.boilerplate/compare/v0.1.0...HEAD